Tweak checker logic for input steps not in steplist
When handling the case of an input step not in the steplist in
_check_flowgraph_io(), we need to check the output dir of the
input step when the input step is not in the steplist, rather than the
inputs dir of the step itself. This is due to a sequencing subtlety: on
the server, the files will not yet be copied into the inputs directory
of the 'syn' step when the first static check is called, but they will
be in the outputs directory of 'import'.